requirements.txt
streamlit
pandas
numpy
python-dotenv
google-generativeai